+++ src/caproj/data/base.py
@@ -1,40 +1,62 @@
-import os
-import sys
+"""
+caproj.data.base
+~~~~~~~~~~~~~~~~
+
+This module contains the core :mod:`caproj.data` read and write functionality
+
+**Module classes:**
+
+.. autosummary::
+
+   BaseData
+
+|
+"""
+
 import logging
+import os
 
 import pandas as pd
 
+from caproj.logger import logfunc
 
-shandler = logging.StreamHandler(sys.stdout)
-formatter = logging.Formatter('%(levelname)s: %(message)s')
-shandler.setFormatter(formatter)
-logger = logging.getLogger()
-logger.setLevel(logging.INFO)
-logger.addHandler(shandler)
+# TODO: move getLogger call inside of ``BaseData`` class
+log = logging.getLogger(__name__)
 
 
 class BaseData(object):
-    """Manage base read/write operations for ``caproj.data`` module classes
+    """Manage base read/write operations for :py:mod:`caproj.data` module classes
 
     :cvar df: pandas.DataFrame working copy either read in from
               from file or from an existing object by using ``BaseData``
-              initializing class methods :meth: `~from_file` or
-              :meth: `~from_object`
+              initializing class methods :meth:`BaseData.from_file` or
+              :meth:`BaseData.from_object`
     :cvar input_df: pandas.DataFrame original input copy, not operated
                     upon by any class methods, and only created if
                     ``copy_input`` parameter set to ``True`` during
-                    ``BaseData`` :meth: `~from_file` or :meth: `~from_object`
+                    :meth:`BaseData.from_file` or :meth:`~BaseData.from_object`
                     class creation
+
+    **Class methods:**
+
+    .. autosummary::
+
+       BaseData.from_file
+       BaseData.from_object
+       BaseData.to_file
+       BaseData.log_record_count
     """
 
     def __init__(self, input_df, copy_input):
         if copy_input:
             self.input_df = input_df.copy()  # input df persists for reference
         self.df = input_df  # all basedata changes applied to this df
+        self.log_record_count()
 
     @classmethod
+    @logfunc(log=log, funcname=True, docdescr=True, argvals=True, runtime=False)
     def from_file(cls, filename, copy_input=False, **read_kwargs):
-        """Invoke ``BaseData`` and read csv or excel into pandas.DataFrame
+        """Invoke BaseData class and read csv into pandas.DataFrame
 
         :param filename: str filename of .csv file to be read
         :param copy_input: bool to specify whether self.input_df persists
@@ -53,8 +75,9 @@ class BaseData(object):
         return cls(input_df, copy_input)
 
     @classmethod
+    @logfunc(log=log, funcname=True, docdescr=True, argvals=False, runtime=False)
     def from_object(cls, input_object, copy_input=False):
-        """Invoke ``BaseData`` and read dataframe from in-memory object
+        """Invoke BaseData and read dataframe from in-memory object
 
         Input objects can be either (a) an existing ``BaseData`` object, in
         which case the ``pandas.DataFrame`` stored within that object will be
@@ -74,7 +97,6 @@ class BaseData(object):
             try:
                 if isinstance(input_object.df, pd.DataFrame):
                     input_df = input_object.df.copy()
-                    # TODO implement self.log = input_object.log.copy()
             except:
                 raise TypeError(
                     'input_object must be either pandas.DataFrame or '
@@ -83,8 +105,9 @@ class BaseData(object):
                 )
         return cls(input_df, copy_input)
 
+    @logfunc(log=log, funcname=True, docdescr=True, argvals=True, runtime=False)
     def to_file(self, target_filename, **to_csv_kwargs):
-        """Save current version of ``self.df`` to file in .csv format
+        """Save current version of BaseData.df to file in .csv format
 
         :param target_filename: str filename to which csv should be written
         :param to_csv_kwargs: optional args to pandas.DataFrame.to_csv()
@@ -93,11 +116,11 @@ class BaseData(object):
         self.df.to_csv(target_filename, index=False, **to_csv_kwargs)
 
     def log_record_count(self, id_col='PID'):
-        """Log number of records and unique projects in :attr:`df`
+        """Log number of records and unique projects in `BaseData.df`
         """
-        logging.info(
+        log.info(
             'Number of project change records: {}'.format(len(self.df))
         )
-        logging.info(
+        log.info(
             'Number of unique projects in dataset: {}'.format(self.df[id_col].nunique())
         )