+++ CHANGELOG.rst
@@ -9,12 +9,11 @@ Future releases
     * Recreate data cleansing pipeline with re-evaluated missing data handling
     * Recreate data sub-setting module
 
-v0.1.0 (XXXX-XX-XX)
+v0.1.0 (Future Release)
 -------------------
 
 * Add data cleansing pipeline as ``data`` submodules
 * Add basic EDA plotting functions to ``visualization`` submodules
-* Add corresponding unit tests
 
 .. todo::
 
@@ -27,6 +26,18 @@ v0.1.0 (XXXX-XX-XX)
         * Add record unique ID column
         * Fill missing project attribute fields based on prior or next earliest available info
         * Fill project budget and schedule values
+
+
+v0.0.3 (2020-06-09)
+-------------------
+
+* Add logging configuration and basic ``caproj.logger`` module functionality
+
+
+v0.0.2 (2020-05-26)
+-------------------
+
+* fix setup.cfg to prevent travis-ci build fail
 
 
 v0.0.1 (2020-05-25)